+++ src/_utils/interfaces.ts
@@ -11,7 +11,6 @@ export interface CommonFieldsProps {
 export interface A11yProps {
   id?: string
   role?: string
-  title?: string
   'aria-label'?: string
   'aria-labelledby'?: string
   'aria-describedby'?: string
@@ -25,7 +24,6 @@ export function pickA11yProps<T extends A11yProps>(source: T): Pick<T, A11yKeys>
   const keys: A11yKeys[] = [
     'id',
     'role',
-    'title',
     'aria-label',
     'aria-labelledby',
     'aria-describedby',