+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- **[FIX]** Remove `title` attribute from `A11yProps`
 - [...]
 
 # v41.10.0 (30/11/2020)